Changes for compatibility with Redmine 5.x / Rails 6.x
